+++ KittyClaw.Web/Components/Pages/CostDisplayTotal.cs
@@ -1,0 +1,7 @@
+namespace KittyClaw.Web.Components.Pages;
+
+public static class CostDisplayTotal
+{
+    public static decimal Sum(IEnumerable<decimal> costs) =>
+        costs.Sum(cost => decimal.Round(cost, 2));
+}